vaccine: extract vaccine related functions from center_info to avoid circular import
