+++ tests/test_center_info.py
@@ -5,9 +5,9 @@ from scraper.pattern.scraper_request import ScraperRequest
 from scraper.pattern.scraper_result import ScraperResult, DRUG_STORE
 from utils.vmd_utils import format_phone_number
 from .utils import mock_datetime_now
-from scraper.pattern.center_info import CenterInfo, convert_csv_address, Vaccine, convert_csv_business_hours, \
-    convert_ordoclic_to_center_info, convert_csv_data_to_center_info, get_vaccine_name, \
-    get_vaccine_astrazeneca_minus_55_edgecase, get_last_scans
+from scraper.pattern.center_info import CenterInfo, convert_csv_address, convert_csv_business_hours, \
+    convert_ordoclic_to_center_info, convert_csv_data_to_center_info, get_last_scans
+from scraper.pattern.vaccine import Vaccine, get_vaccine_name, get_vaccine_astrazeneca_minus_55_edgecase
 
 
 def test_center_info_fill():
