+++ tests/test_scraper.py
@@ -2,7 +2,8 @@ import datetime as dt
 import json
 
 from scraper.export.export_merge import export_data
-from scraper.pattern.center_info import CenterInfo, Vaccine, dict_to_center_info, get_vaccine_name
+from scraper.pattern.center_info import CenterInfo, dict_to_center_info
+from scraper.pattern.vaccine import Vaccine, get_vaccine_name
 from scraper.pattern.scraper_result import GENERAL_PRACTITIONER, ScraperResult
 from utils.vmd_utils import departementUtils
 from scraper.scraper import fetch_centre_slots, get_start_date, gouv_centre_iterator, ialternate
